requirements.txt
google-api-python-client>=2.100.0
requests>=2.31.0
python-dotenv>=1.0.0
pycryptodome>=3.19.0
